pyproject.toml
[tool.poetry]
name = "rest-api"
version = "0.1.0"
description = "rest api conception using FastAPI"
authors = ["KondratovaLudmila <137877876+KondratovaLudmila@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
sqlalchemy = "^2.0.25"
psycopg2 = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.5.3"}
uvicorn = {extras = ["standart"], version = "^0.25.0"}
alembic = "^1.13.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
pydantic-settings = "^2.1.0"
fastapi-mail = "^1.4.1"
cloudinary = "^1.38.0"
fastapi-limiter = "^0.1.6"
redis = "^5.0.1"
pytest = "^8.0.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"


[tool.poetry.group.test.dependencies]
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
